% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellColocation.R
\name{cellColocation}
\alias{cellColocation}
\title{Get the co-localized cell type pairs according to the weights for each pixel
after RCTD pipeline}
\usage{
cellColocation(
  rctd,
  rctd_mode = "multi",
  method = c("pcc", "jac"),
  topn = 10,
  pcc = 0.05,
  pval = 0.05,
  padj = 0.05,
  jac = 0.05
)
}
\arguments{
\item{rctd}{An RCTD object or a data.frame containing the confident results
of the RCTD algorithm.}

\item{rctd_mode}{RCTD analysis, such as in 'multi' mode, 'doublet' mode,
or 'full' mode.}

\item{method}{The method to identify co-localized celltypes, including 'pcc'
(calculating Pearson Correlation Coefficient) and 'jac'
(calculating Jaccard similarity coefficient)}

\item{topn}{Threshold of the number of cell-type pairs to keep according to
the method of identifying co-localized cell type pairs}

\item{pcc}{Threshold of the Pearson Correlation Coefficient for identifying
significant co-localized cell type pairs}

\item{pval}{Threshold of the p-value for identifying significant
co-localized cell type pairs. Only for 'pcc' method}

\item{padj}{Threshold of the ajusted p-value for identifying significant
co-localized cell type pairs. Only for 'pcc' method}

\item{jac}{Threshold of the Jaccard similarity coefficient for identifying
significant co-localized cell type pairs}
}
\value{
a data.frame contains co-localized cell-type pairs
}
\description{
Get the co-localized cell type pairs according to the weights for each pixel
after RCTD pipeline
}
