% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rctd4weights.R
\name{rctd4weight}
\alias{rctd4weight}
\title{Tidy RCTD results}
\usage{
rctd4weight(rctd, rctd_mode = c("multi", "full", "doublet"), conf = TRUE)
}
\arguments{
\item{rctd}{an RCTD object containing the results of the RCTD algorithm.
documentation for more information on interpreting the content of the
RCTD object}

\item{rctd_mode}{the mode of RCTD analysis, such as in 'multi' mode,
'doublet' mode, or 'full' mode}

\item{conf}{whether only return the confident weights when user choose '
multi' mode}
}
\value{
a data.frame of cell type weights for every pixel
}
\description{
Tidy RCTD results
}
