% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_comm.R
\name{st_comm}
\alias{st_comm}
\title{Get the significant cell-cell communication restricted by ligand-receptor
co-expressed in co-localized cell type pairs}
\usage{
st_comm(
  object,
  weights.df,
  ctpairs,
  cellchat,
  db = c("mouse", "human"),
  fisher.pavl = 0.05,
  fisher.padj = 0.05
)
}
\arguments{
\item{object}{Seurat object}

\item{weights.df}{a data frame of cell type confident weights for every pixel}

\item{ctpairs}{a data frame contains co-localized cell-type pairs}

\item{cellchat}{CellChat object or a net data.frame with single-cell data,
which contained the communication probability and infered cellular
communication network}

\item{db}{the species to use for CellChatDB to choose, including 'mouse' and
'human'}

\item{fisher.pavl}{threshold of the p-value for fisher exact test}

\item{fisher.padj}{threshold of the ajusted p-value for fisher exact test}
}
\value{
a data frame contains the significant cell-cell communication
restricted by ligand-receptor co-expressed in co-localized cell type pairs
}
\description{
Get the significant cell-cell communication restricted by ligand-receptor
co-expressed in co-localized cell type pairs
}
